Don't leave dangling TCP sockets after dialing
This also handles some of the race conditions associated with calling
Close twice and/or calling Close at the same time as the remote closes
the link. There are still some race conditions that can happen, which
should be fixed with proper goroutine synchronization through mutexes
and/or channels. Ref issue #28.