Handle inherit race between data and control conns
Since the control and data exchange is handled on different sockets,
there is an inherit race between the two. It is not possible to
guarantee that this won't happen, but waiting a couple of seconds for
the two to "sync up" where it really matters reduces the probability of
bugs due to these race conditions.